Insert time series data values to ODM2 database

Insert time series data values to ODM2 database

db_insert_results_ts(db, datavalues, method, site_code, variables,
  sampledmedium, processinglevel = "Raw data", actionby = NULL,
  equipment = NULL, aggregationstatisticcv = NULL, zlocation = NULL,
  zlocationunits = NULL, ...)

Arguments

db
database connection

datavalues
data frame with columns "Timestamp" with POSIXct YYYY-MM-DD H:M:S format, and column names corresponding to variable names

method
code for method used to collect data

site_code
sampling feature code at which data were collected

variables
a named list defining units for variable names in datavalues data frame with format list("variablen1ame" = "units1name", "variable2name" = "units2name")

sampledmedium
term from controlled vocabulary for medium sampled eg. Air, Water, Soil

processinglevel
code for processing level. will be added to processinglevels table if new. defaults to "Raw data'.

actionby
(optional) the person who performed the action

equipment
(optional) the equipment used to collect the data

aggregationstatisticcv
term from controlled vocabulary

zlocation
(optional) z location offset

zlocationunits
(optional but required if zlocation is set) name of units of z location offset

...
parameters to pass to various db_describe_ functions

Value

true if successful

Examples

# NOT RUN {
db <- rodm2::create_sqlite(connect = TRUE)
tsrv <- data.frame(Timestamp = "2018-06-27 13:55:00",
"Wind direction" = 180, "Wind speed" = 1, "Wind gust speed" = 2)
db_insert_results_ts(db = db, datavalues = tsrv,
method = "SonicAnemometer", site_code = "BB2",
variables = list("Wind direction" = "Degree",
"Wind speed" = "Meter per Second", "Wind gust speed" = "Meter per Second"),
processinglevel = "Raw data",
sampledmedium = "Air")
# }